refactor: move common parts from (V8Profile|V8HoleFuzzingProfile|V8SandboxProfile).swift to V8CommonProfile.swift
